Prevent wrapping within badges, pills and kbd's
This does have the disadvantage that if one of these is too long to fit
on a single line then it'll overflow and break the layout rather than
wrapping, so you can't put long text inside a badge, pill or kbd.

For this reason I think it's best to allow wrapping within other
elements like <code>, <samp>, etc.