Remove the CSS class `lead`
Style `<hgroup>` instead.

This is slightly less flexible: it can only be used for subheadings or
other additional or secondary content that follows a heading in an
`<hgroup>`. You can't just use it to style paragraphs and other elements
anywhere in a document.

But I prefer this approach of styling standard HTML elements instead of
CSS classes. It provides a clearer guide for when and when not to use
this style and what to use it for.