Add more color variables
Add a few missing variables for customizing the color scheme: whenever a
background color is set there should be a variable for the corresponding
foreground color, even if the default color scheme sets it to the same
as the primary foreground color.

Also rename some of the variables, giving them meaningful rather than
numbered names.