Remove the randomization of the banner image
This is causing a noticeable delay between the rest of the page
rendering and then the banner image appearing, even with prefetching.
Just use the `oatcakes.svg` image (which I think is my favourite one)
all the time.

I'm keeping the other two images in the git repo in case I want them,
but they're unused.