Style subheadings only with `<h1>` or `<h2>`
Style a `<p>` as a subheading only if it shares an `<hgroup>` with an
`<h1>` or an `<h2>` and not, for example, if the `<hgroup>` contains an
`<h3`> instead.

Only `<h1>`'s and `<h2>`'s are actually bigger than the subheading style,
so the style only makes sense with those.